+++ WerewolvesCompany/Roles.cs
@@ -35,6 +35,7 @@ namespace WerewolvesCompany
             roles.Add(new Witch());
             roles.Add(new Seer());
             roles.Add(new WildBoy());
+            roles.Add(new Cupid());
 
             return roles;
         }
@@ -105,20 +106,21 @@ namespace WerewolvesCompany
 
         public RolesManager rolesManager = Utils.GetRolesManager();
 
-        public virtual string roleName { get; }
+        public virtual string roleName { get; set; } = "Default Role Name";
 #nullable enable
-        public virtual string? roleNameColor => null;
+        public virtual string? roleNameColor { get; set; }  = null;
 #nullable disable
         public string roleNameColored => GetRoleNameColored();
         public string terminalName => roleName.Replace(" ", "_");
-        public virtual int refInt { get; }
-        public virtual string winCondition { get; }
-        public virtual string roleShortDescription { get; set; }
-        public virtual string roleDescription { get; set; }
+        public virtual int refInt { get; set; } = -1;
+        public virtual string winCondition { get; set; } = "Role Win Condition";
+        public virtual string roleShortDescription { get; set; } = "Role Short Description";
+        public virtual string roleDescription { get; set; } = "Role Description";
         public virtual string rolePopUp  => $"{winCondition} {roleShortDescription}";
         public virtual Sprite roleIcon => null; // Default icon (null if none)
 
-        
+
+
 
         // Who's in range
         public ulong? targetInRangeId { get; set; }
@@ -128,17 +130,26 @@ namespace WerewolvesCompany
 
 
         // ToolTip
-        
-        public virtual string mainActionKey { get { return "Z"; } }
-        public virtual string secondaryActionKey { get { return "V"; } }
-        public virtual string mainActionName { get { return "Main Action Name"; } }
-        public virtual string secondaryActionName { get { return "Secondary Action Name"; } }
 
-        public virtual string mainActionText { get { return $"{mainActionName} {targetInRangeName}"; } }
-        public virtual string mainActionTooltip { get { return $"[{mainActionKey}] {mainActionText} {GetCurrentMainActionCooldownText()}".Trim(); } }
-        public virtual string secondaryActionText { get { return $"{secondaryActionName} {targetInRangeName}"; } }
-        public virtual string secondaryActionTooltip { get { return $"[{secondaryActionKey}] {secondaryActionText} {GetCurrentSecondaryActionCooldownText()}".Trim(); } }
-        public virtual string roleActionText { get { return GetRoleActionText(); } }
+        public virtual string mainActionKey { get; set; } = "Z";
+        public virtual string secondaryActionKey { get; set; } = "V";
+        public virtual string mainActionName { get; set; } = "Main Action Name";
+        public virtual string secondaryActionName { get; set; } = "Secondary Action Name";
+
+        public virtual string mainActionText => $"{mainActionName} {targetInRangeName}";
+        //public virtual string mainActionTooltip { get { return $"[{mainActionKey}] {mainActionText} {GetCurrentMainActionCooldownText()}".Trim(); } }
+        public virtual string mainActionTooltip => GetMainActionTooltip();
+
+        public virtual string secondaryActionText => $"{secondaryActionName} {targetInRangeName}";
+        //public virtual string secondaryActionTooltip { get { return $"[{secondaryActionKey}] {secondaryActionText} {GetCurrentSecondaryActionCooldownText()}".Trim(); } }
+        public virtual string secondaryActionTooltip => GetSecondaryActionTooltip();
+        public virtual string roleActionText
+{
+            get { return GetRoleActionText(); } 
+            set { roleActionText = value; }
+        }
+
+
 
 
 
@@ -149,11 +160,11 @@ namespace WerewolvesCompany
 
 
         // Cooldowns
-        public float currentMainActionCooldown;
-        public float currentSecondaryActionCooldown;
+        public virtual float currentMainActionCooldown { get; set; }
+        public virtual float currentSecondaryActionCooldown { get; set; }
 
-        public float baseMainActionCooldown;
-        public float baseSecondaryActionCooldown;
+        public virtual float baseMainActionCooldown { get; set; }
+        public virtual float baseSecondaryActionCooldown { get; set; }
 
         public bool IsMainActionOnCooldown => (currentMainActionCooldown > 0);
         public bool IsSecondaryActionOnCooldown => (currentSecondaryActionCooldown > 0);
@@ -161,9 +172,15 @@ namespace WerewolvesCompany
 
         // Interactions with others roles
         public bool isImmune = false;
+        public bool amIRomanced = false;
+        public ulong? isInLoveWith = null;
 
-        
         public Role()
+        {
+            InitiateCooldowns();
+        }
+        
+        public virtual void InitiateCooldowns()
         {
             baseMainActionCooldown = baseActionCooldown.Value;
             baseSecondaryActionCooldown = baseActionCooldown.Value;
@@ -171,8 +188,6 @@ namespace WerewolvesCompany
             currentMainActionCooldown = startOfRoundActionCooldown.Value;
             currentSecondaryActionCooldown = startOfRoundActionCooldown.Value;
         }
-        
-        
 
         public string GetRoleActionText()
         {
@@ -195,19 +210,30 @@ namespace WerewolvesCompany
         public void DisplayRolePopUp()
         {
             logdebug.LogInfo("Display the role PopUp");
-            HUDManager.Instance.DisplayTip($"You are a {roleName}", rolePopUp); 
+            HUDManager.Instance.DisplayTip($"{roleName}", rolePopUp); 
             //HUDManager.Instance.DisplayTip($"Test <color=red>red</color>", "Test <color=blue>blue</color>");
         }
 
         public virtual bool IsLocallyAllowedToPerformMainAction()
         {
-            return (!IsMainActionOnCooldown && !(targetInRangeId == null));
+            return (!IsMainActionOnCooldown && !(targetInRangeId == null) && IsLocallyAllowedToPerformMainActionRoleSpecific());
         }
 
         public virtual bool IsLocallyAllowedToPerformSecondaryAction()
         {
-            return (!IsSecondaryActionOnCooldown && !(targetInRangeId == null));
+            return (!IsSecondaryActionOnCooldown && !(targetInRangeId == null) && IsLocallyAllowedToPerformMainActionRoleSpecific());
         }
+
+        public virtual bool IsLocallyAllowedToPerformMainActionRoleSpecific()
+        {
+            return true;
+        }
+
+        public virtual bool IsLocallyAllowedToPerformSecondaryActionRoleSpecific()
+        {
+            return true;
+        }
+
 
         public virtual bool IsAllowedToPerformMainAction()
         {
@@ -283,13 +309,14 @@ namespace WerewolvesCompany
 
 
 
-        public void SetMainActionOnCooldown()
+        public virtual void SetMainActionOnCooldown()
         {
             currentMainActionCooldown = baseMainActionCooldown;
         }
 
-        public void SetSecondaryActionOnCooldown()
+        public virtual void SetSecondaryActionOnCooldown()
         {
+            logdebug.LogInfo("Setting my secondary action on cooldown");
             currentSecondaryActionCooldown = baseSecondaryActionCooldown;
         }
 
@@ -321,42 +348,77 @@ namespace WerewolvesCompany
         }
         //
 
-
-        // Notifications of success and failed actions
-        public virtual void NotifyMainActionSuccess(string targetPlayerName)
+        private string GetMainActionTooltip()
         {
-            HUDManager.Instance.DisplayTip(roleName, "Main action success");
+            if (currentMainActionCooldown < 5000)
+            {
+                return $"[{mainActionKey}] {mainActionText} {GetCurrentMainActionCooldownText()}".Trim();
+            }
+            else
+            {
+                return "";
+            }
+        }
+
+        private string GetSecondaryActionTooltip()
+        {
+            if (currentSecondaryActionCooldown < 5000)
+            {
+                return $"[{secondaryActionKey}] {secondaryActionText} {GetCurrentSecondaryActionCooldownText()}".Trim();
+            }
+            else
+            {
+                return "";
+            }
         }
 
 
+
+        // ------ Main action success
+        //public virtual void NotifyMainActionSuccess(string targetPlayerName)
+        //{
+        //    HUDManager.Instance.DisplayTip(roleName, "Main action success");
+        //}
+        // Alternative parameters inputs for the Seer
+        public virtual void NotifyMainActionSuccess(string targetPlayerName, Role role)
+        {
+            return;
+        }
+        // Alternative parameters inputs for the Wild Boy
+        public virtual void NotifyMainActionSuccess(ulong targetId)
+        {
+            return;
+        }
+
+
+        // ------ Secondary action success
         public virtual void NotifySecondaryActionSuccess(string targetPlayerName)
         {
             HUDManager.Instance.DisplayTip(roleName, "Secondary action success");
         }
+        public virtual void NotifySecondaryActionSuccess(ulong targetId)
+        {
+            HUDManager.Instance.DisplayTip(roleName, "Secondary action success");
+        }
+        public virtual void NotifySecondaryActionSuccess()
+        {
+            HUDManager.Instance.DisplayTip(roleName, "Secondary action success");
+        }
 
+
+        // ------ Main action failed
         public virtual void NotifyMainActionFailed(string targetPlayerName)
         {
             HUDManager.Instance.DisplayTip(roleName, "Main action failed");
         }
 
 
+        // ------ Secondary action failed
         public virtual void NotifySecondaryActionFailed(string targetPlayerName)
         {
             HUDManager.Instance.DisplayTip(roleName, "Secondary action failed");
         }
 
-
-        // Alternative parameters inputs for the Seer
-        public virtual void NotifyMainActionSuccess(string targetPlayerName, Role role)
-        {
-            return;
-        }
-
-        // Alternative parameters inputs for the Wild Boy
-        public virtual void NotifyMainActionSuccess(ulong targetId)
-        {
-            return;
-        }
 
 
         public ulong GrabTargetPlayer()
@@ -372,15 +434,18 @@ namespace WerewolvesCompany
 
     // ----------------------------------------
     // Roles
+
+
+    // Role: Werewolf
     class Werewolf : Role
     {
-        public override string roleName => "Werewolf";
-        public override int refInt => 0;
-        public override string roleNameColor => "red";
-        public override string winCondition => "You win by killing all Villagers";
-        public override string roleShortDescription => "You have the ability to kill other players";
-        public override string mainActionName => "Kill";
-        public override string roleDescription => "The Werewolves shall kill other players before ship departure.\nThe Werewolf has the ability to kill another player.";
+        public override string roleName { get; set; } = "Werewolf";
+        public override int refInt { get; set; }  = 0;
+        public override string roleNameColor { get; set; } = "red";
+        public override string winCondition { get; set; } = "You win by killing all Villagers";
+        public override string roleShortDescription { get; set; } = "You have the ability to kill other players";
+        public override string mainActionName { get; set; } = "Kill";
+        public override string roleDescription { get; set; } = "The Werewolves shall kill other players before ship departure.\nThe Werewolf has the ability to kill another player.";
 
         // Parameters
         public override NetworkVariable<float> interactRange => rolesManager.WerewolfInteractRange;
@@ -401,8 +466,10 @@ namespace WerewolvesCompany
             rolesManager.WerewolfKillPlayerServerRpc(targetId);
         }
 
-        public override void NotifyMainActionSuccess(string targetPlayerName)
+        public override void NotifyMainActionSuccess(ulong targetId)
         {
+            string targetPlayerName = rolesManager.GetPlayerById(targetId).playerUsername;
+
             logger.LogInfo($"Successfully killed {targetPlayerName}.");
             HUDManager.Instance.DisplayTip($"{roleName}", $"You killed {targetPlayerName}.");
         }
@@ -415,14 +482,16 @@ namespace WerewolvesCompany
     }
 
 
+
+    // Role: Villager
     class Villager : Role
     {
-        public override string roleName => "Villager";
-        public override int refInt => 1;
-        public override string winCondition => "You win by killing the Werewolves.";
-        public override string roleShortDescription => "You do not have any special ability.";
-        public override string roleActionText => "";
-        public override string roleDescription => "The Villager shall find and kill the Werewolves before ship departure.\nThe Villager can patpat others players.";
+        public override string roleName { get; set; } = "Villager";
+        public override int refInt { get; set; } = 1;
+        public override string winCondition { get; set; } = "You win by killing the Werewolves.";
+        public override string roleShortDescription { get; set; } = "You do not have any special ability.";
+        public override string roleActionText { get; set; } = "";
+        public override string roleDescription { get; set; } = "The Villager shall find and kill the Werewolves before ship departure.\nThe Villager can patpat others players.";
 
         // Parameters
         public override NetworkVariable<float> interactRange => rolesManager.VillagerInteractRange;
@@ -439,15 +508,17 @@ namespace WerewolvesCompany
     }
 
 
+
+    // Role: Witch
     class Witch : Role
     {
-        public override string roleName => "Witch";
-        public override int refInt => 2;
-        public override string winCondition => "You win by killing the Werewolves.";
-        public override string roleShortDescription => "You have the ability to protect one player, and kill another one.";
-        public override string mainActionName => "Poison";
-        public override string secondaryActionName => "Protect";
-        public override string roleDescription => "The Witch is part of the village. She shall find and kill the Werewolves before ship departure.\nThe Witch has two potions, and can do two things:\n- Poison another player and kill him (once per round)\n- Protect another player and make him immune once to a Werewolf attack (once per round). The immune player won't know he has been immunized, nor will he know he loses his immune status. The Witch cannot protect herself.";
+        public override string roleName { get; set; } = "Witch";
+        public override int refInt { get; set; } = 2;
+        public override string winCondition { get; set; } = "You win by killing the Werewolves.";
+        public override string roleShortDescription { get; set; } = "You have the ability to protect one player, and kill another one.";
+        public override string mainActionName { get; set; } = "Poison";
+        public override string secondaryActionName { get; set; } = "Protect";
+        public override string roleDescription { get; set; } = "The Witch is part of the village. She shall find and kill the Werewolves before ship departure.\nThe Witch has two potions, and can do two things:\n- Poison another player and kill him (once per round)\n- Protect another player and make him immune once to a Werewolf attack (once per round). The immune player won't know he has been immunized, nor will he know he loses his immune status. The Witch cannot protect herself.";
 
         // Parameters
         public override NetworkVariable<float> interactRange => rolesManager.WitchInteractRange;
@@ -475,8 +546,10 @@ namespace WerewolvesCompany
             rolesManager.WitchImmunizePlayerServerRpc(targetId);
         }
 
-        public override void NotifyMainActionSuccess(string targetPlayerName)
+        public override void NotifyMainActionSuccess(ulong targetId)
         {
+            string targetPlayerName = rolesManager.GetPlayerById(targetId).playerUsername;
+
             logger.LogInfo($"Successfully poisoned {targetPlayerName}.");
             HUDManager.Instance.DisplayTip($"{roleName}", $"You poisoned {targetPlayerName}.");
         }
@@ -489,15 +562,17 @@ namespace WerewolvesCompany
     }
 
 
+
+    // Role: Seer
     class Seer : Role
     {
-        public override string roleName => "Seer";
-        public override int refInt => 3;
-        public override string winCondition => "You win by killing the Werewolves.";
-        public override string roleShortDescription => "You have the ability to see a player's role.";
-        public override string mainActionName => "Seer role";
+        public override string roleName { get; set; } = "Seer";
+        public override int refInt { get; set; } = 3;
+        public override string winCondition { get; set; } = "You win by killing the Werewolves.";
+        public override string roleShortDescription { get; set; } = "You have the ability to see a player's role.";
+        public override string mainActionName { get; set; } = "Seer role";
         public override string mainActionText => $"Seer {targetInRangeName}'s role";
-        public override string roleDescription => "The Seer is part of the village. She shall find and kill the Werewolves before ship departure.\nThe Seer can seer another player's role.";
+        public override string roleDescription { get; set; } = "The Seer is part of the village. She shall find and kill the Werewolves before ship departure.\nThe Seer can seer another player's role.";
 
         // Parameters
         public override NetworkVariable<float> interactRange => rolesManager.SeerInteractRange;
@@ -524,19 +599,21 @@ namespace WerewolvesCompany
     }
 
 
+
+    // Role: Wild Boy
     class WildBoy : Role
     {
-        public override string roleName => "Wild Boy";
-        public override int refInt => 4;
-        public override string winCondition => "For now, you win with the village.";
+        public override string roleName { get; set; } = "Wild Boy";
+        public override int refInt { get; set; } = 4;
+        public override string winCondition { get; set; } = "For now, you win with the village.";
         public string _roleShortDescription = "You can idolize a player. If he dies, you become a werewolf.";
         public override string roleShortDescription
         {
             get { return _roleShortDescription; }
             set { _roleShortDescription = value; }
         }
-        public override string roleDescription => "The Wild Boy wins either with the Villagers or the Werewolves, depending on his status.\nThe Wild Boy can target a player who becomes his idol. If his idol dies, the Wild Boy becomes a Werewolf. As long as his idol is alive, he wins with the Villagers.";
-        public override string mainActionName => "Idolize";
+        public override string roleDescription { get; set; } = "The Wild Boy wins either with the Villagers or the Werewolves, depending on his status.\nThe Wild Boy can target a player who becomes his idol. If his idol dies, the Wild Boy becomes a Werewolf. As long as his idol is alive, he wins with the Villagers.";
+        public override string mainActionName { get; set; } = "Idolize";
         public ulong? idolizedId;
 
         // Parameters
@@ -581,4 +658,169 @@ namespace WerewolvesCompany
         }
     }
 
+
+
+    // Role: Cupid
+    class Cupid : Role
+    {
+        public override string roleName { get; set; } = "Cupid";
+        public override int refInt { get; set; } = 5;
+        public override string winCondition { get; set; }  = "You win with the village.";
+        public override string roleShortDescription { get; set; }  = "You can make two players fall in love. They must win together. They also die together.";
+        public override string roleDescription { get; set; } = "Cupid is able to make two players fall deeply in love.\nWhen one lover die, the other one also dies.\nThe lovers win together. If they're part of the same team, they win with their team. If they're part of different teams, they must be the only two survivors.";
+
+        public override string mainActionName { get; set; } = "Romance";
+        public override string secondaryActionName { get; set; } = "Romance Myself";
+        public override string secondaryActionText => secondaryActionName;
+
+        public override float currentSecondaryActionCooldown => currentMainActionCooldown;
+        public override float baseSecondaryActionCooldown => baseMainActionCooldown;
+
+
+        public List<ulong> lovers = new List<ulong>();
+        //public ulong? firstRomancedId;
+        //public ulong? secondRomancedId;
+
+        public bool TargetIsAlreadyRomanced => lovers.Contains(targetInRangeId.Value);
+        public bool AmIAlreadyRomanced => lovers.Contains(Utils.GetLocalPlayerControllerB().OwnerClientId);
+        public bool AreBothTargetsRomanced => (lovers.Count == 2);
+
+        public int romancedPlayersCallbackAmount = 0;
+
+        // Parameters
+        public override NetworkVariable<float> interactRange => rolesManager.CupidInteractRange;
+        public override NetworkVariable<float> baseActionCooldown => rolesManager.CupidActionCoolDown;
+        public override NetworkVariable<float> startOfRoundActionCooldown => rolesManager.CupidStartOfRoundActionCoolDown;
+        
+        
+        public Cupid() : base() { }
+
+        public override void InitiateCooldowns()
+        {
+            baseMainActionCooldown = baseActionCooldown.Value;
+            //baseSecondaryActionCooldown = baseActionCooldown.Value;
+
+            currentMainActionCooldown = startOfRoundActionCooldown.Value;
+            //currentSecondaryActionCooldown = startOfRoundActionCooldown.Value;
+        }
+
+        // Cupid really only has one action: romancing someone. The "secondary" action only serves as the main action targetting himself
+        public override void SetSecondaryActionOnCooldown()
+        {
+            SetMainActionOnCooldown();
+        }
+
+        public override void SetMainActionOnCooldown()
+        {
+            logdebug.LogInfo("Trying to set main action on cooldown");
+            // Only set action on cooldown if two targets have been romanced
+            if (lovers.Count == 2)
+            {
+                logdebug.LogInfo("Success");
+                currentMainActionCooldown = baseMainActionCooldown;
+            }
+        }
+
+
+        public override void PerformMainAction()
+        {
+            logger.LogInfo($"{roleName} is romancing a target: {targetInRangeName}.");
+            rolesManager.CupidRomancePlayerServerRpc(targetInRangeId.Value);
+        }
+
+        public override void PerformSecondaryAction()
+        {
+            logger.LogInfo($"{roleName} is romancing himself.");
+            rolesManager.CupidRomancePlayerServerRpc(Utils.GetLocalPlayerControllerB().OwnerClientId);
+        }
+        
+
+        public override void NotifyMainActionSuccess(ulong targetId)
+        {
+            NotifyRomancingSuccess(targetId);
+        }
+
+        public override void NotifySecondaryActionSuccess(ulong targetId)
+        {
+            NotifyRomancingSuccess(targetId);
+        }
+
+        private void NotifyRomancingSuccess(ulong targetId)
+        {
+            logdebug.LogInfo("I am running the Romancing confirmation Confirmation");
+            lovers.Add(targetId);
+
+            logdebug.LogInfo("Displaying romancing status on HUD");
+            if (!AreBothTargetsRomanced)
+            {
+                logdebug.LogInfo("Only one target is romanced. Notifying Cupid");
+                HUDManager.Instance.DisplayTip($"{roleName}", $"<color=#ff00ffff>{rolesManager.GetPlayerById(targetId).playerUsername}</color> will be romanced.");
+                
+            }
+            else // If both targets are now romanced, we have a couple. Send them that they are now romanced.
+            {
+                logdebug.LogInfo("Both targets are romanced. Sending the lovers their respective lover.");
+                romancedPlayersCallbackAmount = 0;
+                rolesManager.CupidSendLoversTheirLoverServerRpc(lovers[0], lovers[1]);
+            }
+        }
+
+        public void CheckForCallBackOfLover()
+        {
+            logdebug.LogInfo($"Received a callback from a lover. Current value: {romancedPlayersCallbackAmount}");
+            romancedPlayersCallbackAmount += 1;
+            if (romancedPlayersCallbackAmount == 2)
+            {
+                logdebug.LogInfo("Received 2 callbacks from lovers, notifying to Cupid.");
+                string player1Name = rolesManager.GetPlayerById(lovers[0]).playerUsername;
+                string player2Name = rolesManager.GetPlayerById(lovers[0]).playerUsername;
+                roleShortDescription = $"<color=#ff00ffff>{player1Name}</color> and <color=#ff00ffff>{player2Name}</color> are deeply in love. They will die together. They must win together.";
+                if (!lovers.Contains(Utils.GetLocalPlayerControllerB().OwnerClientId))
+                {
+                    DisplayRolePopUp();
+                }
+
+                logger.LogInfo($"Successfully romanced {player1Name} and {player2Name}");
+
+                SetMainActionOnCooldown();
+            }
+        }
+
+        
+
+        public override bool IsLocallyAllowedToPerformMainActionRoleSpecific()
+        {
+            if (TargetIsAlreadyRomanced)
+            {
+                HUDManager.Instance.DisplayTip($"{roleName}", $"{targetInRangeName} is already romanced.");
+                return false;
+            }
+            return true;
+        }
+
+        public override bool IsLocallyAllowedToPerformSecondaryActionRoleSpecific()
+        {
+            if (AreBothTargetsRomanced)
+            {
+                return false;
+            }
+            {
+                
+            }
+            if (AmIAlreadyRomanced)
+            {
+                HUDManager.Instance.DisplayTip($"{roleName}", $"You are already romanced.");
+                return false;
+            }
+            return true;
+        }
+
+        public override bool IsLocallyAllowedToPerformSecondaryAction()
+        {
+            return (!IsSecondaryActionOnCooldown && IsLocallyAllowedToPerformSecondaryActionRoleSpecific());
+        }
+    }
 }
+
+
+