+++ CHANGELOG.md
@@ -6,10 +6,12 @@ Additions
 - Werewolf role is now displayed in red at the top of the screen
 - The seer now see werewolves roles in red
 - Added terminal command to get informations on a specific role : 'wc RoleName'
-
+- For roles that can use their action only once per round, disable the tooltip on aiming at a player after using the action
+ 
 Bug fixes:
 - Fixed a bug introduced in 0.3.0 that prevented adding multiple roles at once ('wc add role1 role2')
-
+- Re-organized the config file sections
+ 
 ### [0.3.1]
 
 Turned off the debugging entries